endpoints + unit tests

+++ app/models/Users.py
@@ -1,7 +1,7 @@
 import bcrypt
 import re
 from abc import ABC, abstractmethod
-from typing import Union, Optional, List, Dict
+from typing import Union, Optional, List, Dict, Tuple
 from typeguard import typechecked
 
 from app.data.DbConnection import SessionLocal, BasicUserDB, UserDB, ManagerDB, OrdersDB
@@ -322,6 +322,8 @@ class Authentication:
             return True
         if isinstance(credit_card_num, int):
             return True
+        if credit_card_num > 9999999:
+            return True
         return False
 
 
@@ -394,7 +396,6 @@ class BasicUser(ABC):
             raise ValueError("Invalid email format")
         return email.lower()
 
-    @abstractmethod
     def set_password(self, new_password: str) -> None:
         """
         Set a new password for the user.
@@ -402,7 +403,8 @@ class BasicUser(ABC):
         Args:
             new_password (str): The new plain text password
         """
-        pass
+        Authentication().set_new_password(self, new_password)
+        return
 
     @abstractmethod
     def __repr__(self) -> str:
@@ -549,15 +551,6 @@ class User(BasicUser):
         """
         return str(self.cart)
 
-    def get_order_hist(self) -> List[Order]:
-        """
-        Retrieve user's order history.
-
-        Returns:
-            List[Order]: List of Order objects associated with the user
-        """
-        return self._orders
-
     def get_order_hist_from_db(self) -> Optional[List[Dict[str, str]]]:
         """
         Gets all user's orders history from the DB.
@@ -584,17 +577,9 @@ class User(BasicUser):
         finally:
             session.close()
 
-    def set_password(self, new_password: str) -> None:
-        """
-        Set a new password for the user.
-
-        Args:
-            new_password (str): The new plain text password
-        """
-        Authentication().set_new_password(self, new_password)
-        return
-
-    def checkout(self, credit_card_num: int, coupon_code: Optional[str] = None) -> bool:
+    def checkout(
+        self, credit_card_num: int, tax: int, coupon_code: Optional[str] = None
+    ) -> Tuple[bool, str]:
         """
         Process checkout of items in the shopping cart.
 
@@ -609,13 +594,15 @@ class User(BasicUser):
             ValueError: If cart is empty, item quantity is invalid, or credit card is invalid
         """
         try:
-            ans = False
             if self.cart.items:
                 for item, quantity in self.cart.items:
                     if not item.check_availability(quantity):
                         raise ValueError(f"There is not enough:{item} in the inventory")
             else:
                 raise ValueError("There are no items in the cart")
+            if (tax is None) or (tax == "None") or (not tax):
+                tax = 0
+            self.cart.apply_tax_on_cart(tax_rate=tax)
 
             if coupon_code:
                 discount_percent, coupon_id = self.cart.get_coupon_discount_and_id(
@@ -645,12 +632,10 @@ class User(BasicUser):
 
             new_order = Order(self.email, self.cart, coupon_id)
             self._orders.append(new_order)
-            ans = True
-
+            self.cart = ShoppingCart()
+            return True, "Checkout successful!"
         except Exception as e:
-            print(f"Error in checkout Proc: {e}")
-        finally:
-            return ans
+            return False, f"Error in checkout process: {e}"
 
     def __repr__(self) -> str:
         """
@@ -727,31 +712,27 @@ class Manager(BasicUser):
         """
         session = SessionLocal()
         try:
-            user_db = session.query(UserDB).filter(UserDB.email == email).first()
             basic_user_db = (
                 session.query(BasicUserDB).filter(BasicUserDB.email == email).first()
             )
+            user_db = session.query(UserDB).filter(UserDB.email == email).first()
+            if not user_db:
+                raise ValueError
             if user_db:
                 session.delete(user_db)
+                session.commit()
+
             if basic_user_db:
                 session.delete(basic_user_db)
+                session.commit()
 
-            session.commit()
+        except ValueError:
+            raise ValueError("Not found")
         except Exception as e:
             session.rollback()
             raise Exception(f"Error deleting user: {e}")
         finally:
             session.close()
-
-    def set_password(self, new_password: str) -> None:
-        """
-        Set a new password for the manager.
-
-        Args:
-            new_password (str): The new plain text password
-        """
-        Authentication().set_new_password(self, new_password)
-        return
 
     def add_manager(self, name: str, email: str, password: str) -> Optional["Manager"]:
         """
@@ -781,7 +762,7 @@ class Manager(BasicUser):
         session = SessionLocal()
         try:
             order_db = session.query(OrdersDB).filter(OrdersDB.id == order_id).first()
-            if order_db.id == OrderStatus.DELIVERED:
+            if order_db.Ostatus == OrderStatus.DELIVERED:
                 raise ValueError
             else:
                 order_db.Ostatus += 1
